requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
colorama==0.4.6
distro==1.9.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
markdown-it-py==3.0.0
mdurl==0.1.2
numpy>=1.20.0
openai>=1.0.0
pydantic==2.10.6
pydantic_core==2.27.2
pyfiglet>=0.8.0
Pygments==2.19.1
python-dotenv>=1.0.0
rich>=13.0.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
hnswlib>=0.7.0
chromadb>=0.4.0
scikit-learn>=1.0.0
beautifulsoup4>=4.10.0
markdown>=3.4.0
PyPDF2>=3.0.0
python-docx>=0.8.11
requests>=2.28.0
pyyaml>=6.0.0
bs4>=0.0.1

# 선택적 의존성
# pinecone-client>=2.2.0  # Pinecone 벡터 데이터베이스 사용 시 필요
# psycopg2-binary>=2.9.0  # pgvector 벡터 데이터베이스 사용 시 필요
# pymupdf>=1.20.0         # PDF 이미지 추출 시 필요
